requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
redis==5.0.1
requests==2.32.3
pycryptodome==3.20.0